Make TS and Rust project both use 2 worker threads
