--- benchmarks/rust-tokio/examples/unfavorable.rs
@@ -1,10 +1,0 @@
-use rust_tokio::{Request, Response};
-
-#[tokio::main]
-async fn main() {
-    rust_tokio::serve(unfavorable).await.unwrap();
-}
-
-pub async fn unfavorable(_request: Request<'_>) -> Response {
-    include_str!("../Cargo.toml").into()
-}
