+++ benchmarks/rust-tokio/src/lib.rs
@@ -1,90 +1,4 @@
-use std::future::Future;
-use std::io::ErrorKind;
-use std::net::{IpAddr, Ipv4Addr, SocketAddr};
-
-use tokio::net::{TcpListener, TcpStream};
-
-mod error;
-mod request;
-mod response;
-
-pub use error::{Error, Result};
-pub use request::{ParseRequestError, Request};
-pub use response::{ContentType, Response, StatusCode};
-
-const HOST: IpAddr = IpAddr::V4(Ipv4Addr::LOCALHOST);
-const PORT: u16 = 8000;
-
-const ADDRESS: SocketAddr = SocketAddr::new(HOST, PORT);
-
-pub async fn serve<F, Fut>(handler: F) -> Result<()>
-where
-    for<'r> F: Fn(Request<'r>) -> Fut,
-    Fut: Future<Output = Response>,
-{
-    let listener = TcpListener::bind(ADDRESS).await?;
-
-    worker(&listener, handler).await?;
-
-    Ok(())
-}
-
-async fn worker<F, Fut>(listener: &TcpListener, mut handler: F) -> Result<()>
-where
-    for<'r> F: FnMut(Request<'r>) -> Fut,
-    Fut: Future<Output = Response>,
-{
-    while let Ok((stream, _addr)) = listener.accept().await {
-        // Use a buffer initialized with zeroes
-        let mut buf = [0; 1024];
-
-        // Fill buffer with request
-        read_request(&stream, &mut buf).await?;
-
-        // Parse request
-        let request = std::str::from_utf8(&buf)?;
-        let request = Request::from_str(request)?;
-
-        // Handle request
-        let response = handler(request).await;
-
-        // Return response
-        write_response(&stream, response.to_string().as_bytes()).await?;
-    }
-
-    Ok(())
-}
-
-async fn read_request(stream: &TcpStream, buf: &mut [u8]) -> Result<()> {
-    let mut bytes_read = 0;
-
-    loop {
-        stream.readable().await?;
-
-        match stream.try_read(&mut buf[bytes_read..]) {
-            Ok(0) => return Ok(()),
-            Ok(n) => bytes_read += n,
-            Err(e) if e.kind() == ErrorKind::WouldBlock => continue,
-            Err(e) => return Err(e.into()),
-        }
-    }
-}
-
-async fn write_response(stream: &TcpStream, response: &[u8]) -> Result<()> {
-    let mut bytes_written = 0;
-
-    loop {
-        stream.writable().await?;
-
-        match stream.try_write(response) {
-            Ok(n) => {
-                bytes_written += n;
-                if bytes_written == response.len() {
-                    return Ok(());
-                }
-            }
-            Err(e) if e.kind() == ErrorKind::WouldBlock => continue,
-            Err(e) => return Err(e.into()),
-        }
-    }
-}
+pub mod error;
+pub mod request;
+pub mod response;
+pub mod server;
